+++ extension/background_entry.js
@@ -1,0 +1,1 @@
+importScripts('background.js', 'background_patch_v0521.js');